Fixes critical login issue due to changed auth URLs. Moves 'model' package to 'dto' to fix checkstyle warnings.
