+++ bundles/org.openhab.binding.iaqualink/src/main/java/org/openhab/binding/iaqualink/internal/api/dto/Device.java
@@ -10,7 +10,7 @@
  *
  * SPDX-License-Identifier: EPL-2.0
  */
-package org.openhab.binding.iaqualink.internal.api.model;
+package org.openhab.binding.iaqualink.internal.api.dto;
 
 /**
  * Device refers to a iAqualink Pool Controller.
